requirements.txt
pandas==1.2.1
fasttext==0.9.2
streamlit==0.75.0
texthero==1.0.9
Pillow==8.1.0
